pyproject.toml
[project]
name = "europa1400-manager"
version = "0.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aenum>=3.1.16",
    "dataclass-wizard>=0.35.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "typer>=0.16.0",
    "tzdata>=2025.2",
]
